sockIO模块结束

+++ sylar/hook.h
@@ -84,7 +84,7 @@ extern close_fun close_f;
 typedef int (*fcntl_fun)(int fildes, int cmd, ...);
 extern fcntl_fun fcntl_f;
 
-typedef int (*ioctl_fun)(int fildes, int request, ... /* arg */);
+typedef int (*ioctl_fun)(int fildes, unsigned long int request, ... /* arg */);
 extern ioctl_fun ioctl_f;
 
 typedef int (*getsockopt_fun)(int socket, int level, int option_name,
